Added delete endpoint, improved template

+++ flask/init_db.py
@@ -1,6 +1,8 @@
 import os
+import os.path
 
 from app import engine, Base
 
-os.remove('test.db')
+if os.path.isfile('test.db'):
+    os.remove('test.db')
 Base.metadata.create_all(engine)
